🎨 chore(lint): scoped clang-tidy target + apply suggested fixes
The previous `make lint` recipe shelled out to clang-tidy with
`-p build` and silently swallowed every error. That made it a
no-op in practice — the lint target couldn't see the project
include paths, and the suppress-on-failure made any real findings
invisible. Make it actually useful.

- .clang-tidy: tighten HeaderFilterRegex to match the absolute
  project path so headers from sibling worktrees stop bleeding
  into the analysis.
- Makefile: introduce LINT_SOURCES with the host-portable C files
  (no ESP-IDF deps) and have `make lint` run clang-tidy directly
  with the test compile flags + include set. Errors now surface
  on stderr.
- Apply the resulting findings across components and main:
  - Replace `uint8_t` loop counters with `size_t` where compared
    against a count of the same width (avoid the implicit
    promotion warning).
  - Brace single-statement bodies in if/for to satisfy
    readability-braces-around-statements.
  - Minor pointer-spacing cleanups picked up alongside.
- main/godzilla_img.h: reflow the embedded byte table to one
  entry per line so future diffs stay readable when patches edit
  individual pixels. Byte content is byte-identical to HEAD
  (verified via md5sum of the extracted hex stream).

+++ components/crt_compose/crt_compose_layers.c
@@ -242,7 +242,7 @@ bool crt_compose_viewport_layer_fetch(void *ctx, uint16_t logical_line, uint8_t 
         return true;
     }
 
-    for (uint16_t i = 0; i < out_count; ++i) {
+    for (uint32_t i = 0; i < out_count; ++i) {
         const uint16_t source_x =
             crt_compose_wrap_i32((int32_t)i + layer->scroll_x, layer->source_width);
         idx_out[layer->viewport_x + i] = layer->scratch[source_x];
